requirements.txt
Django==2.2.5
pytest-django==3.5.1
pytest-cov==2.7.1
pytest-html==1.21.1